refactor: split visible-items filtering out of cursor.go
Move computeVisibleMiddleItems, categoryCounts, syncExpandedGroup, and
the filteredX helpers into a new cursor_filter.go alongside
cursor_visible_items_cache.go. Split computeVisibleMiddleItems into
applyTextFilter, sortItemsByFuzzyScore, and applyGroupCollapse to
bring it under the gocyclo cap.
